requirements.txt
streamlit==1.15.2
torch==1.13.0
torchvision==0.14.0
Pillow==9.3.0
scipy==1.9.3
numpy==1.23.5
seaborn==0.12.1
